refactor: 불필요한 action 제거
