+++ app/actions/fetchExchangeRates.ts
@@ -1,17 +1,11 @@
 'use server';
 
-import fetch from 'node-fetch';
-import { calculateValidDate } from '../utils/calculateValidDate';
-
 export async function fetchExchangeRates() {
     try {
-        const searchDate = calculateValidDate();
-        const API_KEY = process.env.EXCHANGERATE_API_KEY;
-
         const response = await fetch(
-            `https://www.koreaexim.go.kr/site/program/financial/exchangeJSON?authkey=${API_KEY}&searchdate=${searchDate}&data=AP01`,
+            'https://cdn.jsdelivr.net/npm/@fawazahmed0/currency-api@latest/v1/currencies/eur.json',
             {
-                agent: new (require('https').Agent)({ rejectUnauthorized: false }),
+                next: { revalidate: 3600 }, // 1시간마다 갱신
             }
         );
 
@@ -19,17 +13,36 @@ export async function fetchExchangeRates() {
             throw new Error('Failed to fetch exchange rates');
         }
 
-        const data = (await response.json()) as ExchangeRate[];
+        const data = (await response.json()) as ExchangeRateByEuro;
+
+        const krwRate = data.eur.krw; // 1유로당 KRW
+        const usdRate = data.eur.usd; // 1유로당 USD
+
+        if (!krwRate || !usdRate) {
+            throw new Error('Missing KRW or USD rate in the response');
+        }
+
+        // 1% 송금 스프레드 가산
+        const krwWithSpread = krwRate * 1.01;
+
+        // Calculate 1달러당 KRW
+        const usdToKrwRate = krwWithSpread / usdRate;
 
         return {
-            rates: data.filter((rate: ExchangeRate) => ['EUR', 'USD'].includes(rate.cur_unit)),
-            referenceDate: calculateValidDate(true),
+            date: data.date,
+            rates: {
+                eurToKrw: parseFloat(krwWithSpread.toFixed(2)), // 1유로당 한화 금액 (송금 수수료 포함)
+                usdToKrw: parseFloat(usdToKrwRate.toFixed(2)), // 1달러당 한화 금액
+            },
         };
     } catch (error) {
         console.error('Exchange rate fetch error:', error);
         return {
-            rates: [],
-            referenceDate: '날짜 정보 없음',
+            date: '',
+            rates: {
+                eurToKrw: null,
+                usdToKrw: null,
+            },
         };
     }
 }
